+++ mod.ts
@@ -44,6 +44,7 @@ export default function lume(
     .ignore("import_map.json")
     .ignore("deno.json")
     .ignore("deno.jsonc")
+    .data("mergedKeys", { tags: "stringArray" })
     .use(url(pluginOptions.url))
     .use(json(pluginOptions.json))
     .use(markdown(pluginOptions.markdown))