+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Fixed
 - Updated dependencies: `pagefind`, `std`, `terser`, `postcss-import`, `lightningcss`.
 - MDX renderer must return a JSX object instead of a string.
+- Removed hardcode configuration of tags merging.
 - URL of the pages processed by `esbuild` when `splitting` is enabled [#323].
 
 ## [1.13.2] - 2022-11-30